// src/api/observer-v2-schema.ts
export const observerOperations={
  "listObserverSessionsV2": {
    "path": "/v1/collaboration/2/projects/{id}/observer-sessions",
    "method": "POST"
  },
  "issueObserverAttachmentV2": {
    "path": "/v1/collaboration/2/observe-grants/{id}/attachments",
    "method": "POST"
  }
} as const;